+++ arangod/RocksDBEngine/RocksDBReplicationContext.h
@@ -146,7 +146,7 @@ class RocksDBReplicationContext {
     // forwarded methods
     bool ok() const { return _result.ok(); }
     bool fail() const { return _result.fail(); }
-    uint64_t errorNumber() const { return _result.errorNumber(); }
+    int errorNumber() const { return _result.errorNumber(); }
     std::string errorMessage() const { return _result.errorMessage(); }
     bool is(uint64_t code) const { return _result.is(code); }
